Fix import_from_JSON. Move checkSaveAndClear() to figureModel
